fixed data type

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,14 +8,14 @@
 int main(void)
 {
 	int count;
-	int num1, num2, sum;
+	unsigned long num1, num2, sum;
 
 	num1 = 0;
 	num2 = 0;
 	for (count = 0; count < 50; count++)
 	{
 		sum = num1 + num2;
-		printf("%d", sum);
+		printf("%lu", sum);
 		num1 = num2;
 		num2 = sum;
 		if (count == 49)
